requirements.txt
asgiref
Django~=5.1.6
django-cors-headers
djangorestframework~=3.15.2
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv~=1.0.1
requests~=2.32.3
geopy~=2.4.1

